requirements.txt
pymupdf
pillow
spacy
opencv-python
